backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and data processing
pandas==2.1.3
sqlalchemy==2.0.23

# Google Sheets integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Data matching and processing
thefuzz==0.20.0
python-Levenshtein==0.23.0

# Validation and serialization
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
structlog==23.2.0